requirements.txt
pandas>=1.5
numpy>=1.23
scipy>=1.9
scikit-learn>=1.1
statsmodels>=0.13
matplotlib>=3.6
openpyxl>=3.0
pyyaml>=6.0
